added e2e test for companies index page
