+++ public/views/partials/companies/index.html
@@ -1,0 +1,7 @@
+<main ng-controller="companiesCtrl as ctrl">
+  <ul ng-repeat="company in ctrl.companiesList">
+    <li>{{ctrl.company}}
+    </li>
+
+  </ul>
+</main>
