+++ tests/e2e/conf.js
@@ -1,6 +1,6 @@
 exports.config = {
   seleniumAddress: 'http://localhost:4444/wd/hub',
-  specs: ['indexFeatures.js'],
+  specs: ['companiesIndexFeatures.js'],
   capabilities: {
     browserName: 'phantomjs'
   },
