--- tests/e2e/pages/index.page.js
@@ -1,10 +1,0 @@
-var IndexPage = function(){
-  this.header = element(by.className("title"));
-};
-
-
-IndexPage.prototype.get = function(){
-  browser.get("http://localhost:8080");
-};
-
-module.exports = IndexPage;
